refactor(rust): Add 'monthname()' and 'dayname()' methods to 'UTC/LocalDateTime'
You'd use to index into an array in a C-like fashion to convert from
month/day number to month/day name. It still works like this under the
hood, but now there's a layer of abstraction over it, with idiomatic
'Error's instead of juggling with '99' as error value and panics.