refactor(rust)!: Return new 'Phase' enum instead of separate values in 'MoonPhase'
In the same vein as the previous commit, the way phase info was returned
was very much a calque of the C API. Now, instead of three separate
fields we return a single instance of the new 'Phase' enum, which
replaces the raw phase integer. Phase name and icon fields disappear in
favor of the new 'name()' and 'icon()' enum method. They're still part
of the JSON version however.

This commit stabilizes the API, as everything that was too C-like has
now disappeared.